sda: Fix the error on casting void pointer to int
When using clang with options '-Werror,-Wvoid-pointer-to-int-cast',
there is the compile error.
        error: cast to smaller integer type 'int' from 'const void *'

Use the C99 uintptr_t to cast first.

(cherry-picked from commit 8e91756430236875c284caba15227e6f523fa278)
CR-Id: ALPS05539426
Feature: External Memory Interface(emi)

+++ drivers/misc/mediatek/sda/cache-parity.c
@@ -545,7 +545,7 @@ static int cache_parity_probe(struct platform_device *pdev)
 
 	dev_info(&pdev->dev, "driver probed\n");
 
-	cache_parity.ver = (int)of_device_get_match_data(&pdev->dev);
+	cache_parity.ver = (int)(uintptr_t)of_device_get_match_data(&pdev->dev);
 
 	atomic_set(&cache_parity.nr_err, 0);
